feat: mypage api 연결 및 데이터 바인딩 적용

+++ app/src/main/java/com/teamwss/websoso/data/model/RepresentiveAvatarEntity.kt
@@ -1,7 +1,7 @@
 package com.teamwss.websoso.data.model
 
 data class RepresentiveAvatarEntity(
-    val avatarId: Int,
+    val avatarId: Long,
     val avatarLine: String,
     val avatarTag: String,
     val userNickname: String,
